(f19323): [REFACTOR] RingList
Implemented a RingList "widget".  It's fairly minimal at this point;
numerous improvements and extensions to be made as necessary.
Introduced Adapter and BaseAdapter, adapted (heh) for GVRSceneObject
instead of View.

Modified HorizontalPicker as proof of concept for RingList.
Currently aligns the left edge of the left-most item with the camera;
not pretty, but not important with the Settings overhaul in the
works.  Also name-swapped heightQuad and widthQuad.

Miscellaneous other:
- Removed unused constructor from BackgroundManager
- Added removeCallbacks() method to MainThread.  Thought I needed it,
then didn't, but since it's written, might as well have it in since
it is a useful facility.
- Corrected spelling of "Policy.backgroundQuardRho"

Change-Id: I94732d5ac6f7cc212e165e7307393dfa386dad86