(f18795) GVRf Widgets
A second pass at widget infrastructure for VR Top (GVRf?).

Per discussions with Tom and Mihail, Widgets are now not derived from
GVRSceneObject, but have GVRSceneObjects backing them.  This makes
for a much tighter interface.

Goals include:

- Providing a clear place for GL thread initialization to occur
- Integrated and automatic TouchManager support
- Provide line-of-sight focus and long focus support
- Better management of show/hide semantics

The API is roughly based on the Android View API; we'll extend it
further as needed.

Patch 3: Fixed FocusManager and rebased.

Patch 4: Added TextWidget. Modified LinearLayout to support curved
layouts; see javadoc for how this is not the same as a ring layout.
Minor additions to LayoutHelpers.

Patch 5: Pulled the curved layout support out of LinearLayout; while
enabling LinearLayout to curve make sense to me for the relevant use
cases, I think it will be unintuitive to others.  Moved all the
curving layout code into a new layout class, RingLayout, which
supports laying out in a ring around its local origin and around a
"virtual" origin, i.e., one that is `radius` units away from the
actual origin.  Javadoc still explains it more fully.

Patch 6: Fixed missing layout() overload in SceneObjectWidget.

Patch 7: Hiding the GVRTransform.  Ideally, we'd like the layouts to
control everything and not have to directly perform tranformations on
the widgets.  For the time being, however, layouts will be able to
organize widgets relative to one another, and not have the full
control that they do in the Android API.

Change-Id: I056c68c4895347b5997de11e1d3aeed647870edd