Update Widgets in preparation for factory creation
All Widgets now have a public constructor for (GVRContext,
GVRSceneObject) for wrapping existing scene objects; these are for
use with a WidgetFactory for loading scenes created with external
tools.

SceneObjectWidget has been renamed AbsoluteLayout and now has a
constructor that takes a width and height.  Use this layout when you
want total control over the layout of child objects.

Also modified GroupWidget:

- Added findChild() by name method
- Added addChildInner() method that adds the child, but doesn't
trigger a layout.  This will be useful for bulk adding children that
are loaded from an external model.

Change-Id: I7ecae9def47db368cbb2e4e178caa8200e25c9d1