Updates to Widget framework
Various updates the the Widget framework and Helpers to support work
on halos, global menu, app panel, etc.

Helpers:
Added various methods to assist in working with GVRSceneObjects in
models.  There is some overlap with VRAssets; we can reconcile later.
These are factored out of ModelContentScene; their removal from there
will be done in a later clean-up commit.

GroupWidget:
- Added methods for adding children by index
- Forwarded one version of removeChild() to another to remove
duplicate code
- Added protected createChild() and createChildren() methods to
assist in constructing GroupWidgets from models
- Now using a List instead of a Map to contain the child Widgets. We
weren't using the Map to do findByName() anyway and there was an
issue with adding unnamed child Widgets: they would overwrite each
other, so you could only ever have one unnamed child Widget.  Names
are intended to be a convenience for looking up Widgets, not a
requirement, and there are definitely instances in which it's handier
to be able to create unnamed Widgets.

Widget:
- Fixed visibility attribute. The loader mashes all attribute names
to lowercase, but the Visibility enum members are all uppercase; the
mismatch resulted in an exception.
- Added method to get the Widget's Android Context.
- Changed how set/getName() works.  Previously, it wrote and read
directly to/from GVRSceneObject.set/getName().  However, since we're
encoding metadata in the scene object's name and there are several
places where we rely on that metadata being intact, it's not a good
idea to overwrite the GVRSceneObject's name.  But for Widget look-up,
we only want to use the actual *name*, not the entire metadata
string.  So now they're separated: get/setName() for the actual name,
and getMetadata() for the metadata.
- Added rotateWithPivot() wrapper method.
- Added proper handling for PLACEHOLDER visibility

WidgetFactory:
- Added several overloads of createWidget():
 - Create with a specified Widget class
 - Create from a named sub-node of a root scene object
 - Create from a named sub-node of a root scene object with a
specified Widget class
 - Create directly from a model file, defaulting to AbsoluteLayout
 - Create directly from a model file with a specified Widget class
 - Create directly from a named sub-node of a model file witha
specified Widget class

Change-Id: Ifad4a3dc69ea31b9c4c5f37281c9cf9c5f4146b4