(b19803,b19823) Clicking wallpaper causes Home overlay to disappear
The fix involves three changes:

1. Fix TouchManger.OnTouch.touch() return value handling
2. Add back key handling support for GVRSceneObject and Widget in
TouchManger
3. Implement touch and back key handling in AppPanelController

The return value from OnTouch.touch() was mostly ignored in
TouchManager.  We had no need for cascaded event handling, so the
value was only checked on interceptors.  As a result, a few places
were returning incorrect values; these were updated to return `true`
when the event cascade should be stopped and `false` to continue it,
as documented for OnTouch.touch().  The interceptor-related code was
checking for `false` to halt cascading; this was corrected.

TouchManager.OnBackKey extends OnTouch to minimize the changes to the
internal logic and public interface.  Otherwise, a lot more code
would have gotten touched.  Widget's added support for back key
parallels its touch event support; making a Widget touchable means it
is eligible for both touch and back key events.  As with touch, the
default handling for back key is to pass it along.

AppPanelController's new handling for the background swallows touch
events and notifies the listener of back key events.
LauncherViewManager uses this to hide the AppPanel as appropriate.

Change-Id: Iece71936fd3fa6e46d7291ec3f6b2e5d4dc27041