(b19821) [UI] App tray location should be in the center of the view
Added set/getBalanceLayout() to RingList.  When set to `true`, items
in the list are laid out equally to the left and right of the
RingList's "center"; the default to is lay all items out to the right
of the "center" (which works nicely in the Settings content scene for
background picker).

Added getLastFocusedAppRotation() to AppRing.  This is used to set
"front" to the center of the most recently focused app. (whereas
previously it was the current camera yaw was).  LauncherViewManager
now updates the front yaw whenever an app gets focused.

AppTray makes use of both of those changes to center itself on the
focused app.

Also added a delay so that the initial set of the front yaw happens
after the splash screen times out; this helps cut down on instances
in which the AppPanel is not directly in front of the user.

Change-Id: Icd0844170510b1503fd43c825a1e26827f713ab6