(b19856,b19787) Apps bleed through the Home Menu screen transparency
The solution here is to give everything in the App Panel --
background, buttons, and labels -- a TRANSPARENT rendering order.
This is not how it should be: it should be possible to set just the
background transparent and leave everything else opaque, so there may
be an issue in GVRf there.

It was also necessary to adjust the render order of
HomeButtonRecentPlane to TRANSPARENT + 1 so that it doesn't "z fight"
with its sibling, HomeButtonRecentIcon.

In the course of experimenting, I tried using GL's offset/offset
factor to solve HomeButtonRecentPlane's issue.  It didn't work, but
it's functionally we may well end up using anyway (there are several
places we use it directly on GVRSceneObject instances) so I left it
in.

I also saw a null pointer exception once that had to do with
attempting to operate on a null GVRRenderData instance.  Not having
render data is a valid state for items with visibility PLACEHOLDER,
so I modified all the places in Widget that work with GVRRenderData
to make null reference checks.

Change-Id: I4e39100dad17b6b193a03ff0ba85eb6c5a50a850