(f19882) Status Bar and Global Menu positioned at "front"
StatusBarController now (mostly) hooked up to model with Widgets

The one exception is the time, which is currently super tiny.  I'll
work with Mike to get a proper sized bit of geometry for it.

Added OnFrontRotationChangedListener interface to MainScene, with
attendant methods.

Modified GlobalMenuController, StatusBarController,
StatusBarBatteryController, and StatusBarWiFiController to make use
of the new interface to listen for changes to "front" and realign in
response.

Also:

- Added exception handling to MainScene.setScale() when calling
listeners
- Added start/stop for StatusBarTimeController
- Fixed an infinite recursion issue with
GroupWidget.findChildByNameInAllGroups().
- Added Enum versions of getProperty() to NodeEntry plus a better
toString() implementation than the default.
- Replaced GVRTextViewSceneObject with TextWidget in NumberPicker.
Did it as a test, but since it works, might as well check it in.
- TextWidget.maybeWrap() will create a GVRTextViewSceneObject
matching the geometry of a host GVRSceneObject if that scene object
isn't already a GVRTextViewSceneObject.

Change-Id: I8d76cceca37f6c8f28b7150e3e12d919729a34cf