(b19883) Animate Status Bar and Global Menu between "front" positions
Changes to implement the Status Bar and Global Menu animating between
AppScenes as the focus changes.

In the main, this consists of modifying
MainScene.OnFrontRotationChangedListener providing the old and new
rotation values to onFrontRotationChanged() and StatusBarController
and GlobalMenuController utilizing the message to run animations.

This change also brings in a set of Widget-based wrappers for the
basic GVRAnimation classes.  They're light-weight and, in addition to
working directly with Widgets, add a number of things that we've had
to hack on top of GVRAnimation: multiple finish listeners, a method
to force the animation to finish, etc.

A few other miscellaneous clean-up changes.

Change-Id: I28649872967ceb5955d9d9ae3c8efd161bdc4bfc