(b19967,b19948) Fix vrScreen callback for onTopTaskIdUpdated
Patch 10: Fixes a compile error in patch 9.

In addition to Mingsong's original changes, this adds a fix for
19948, "Chrome does not re-launch in app ring".
VirtualSpaceScreen.startTask() can throw exceptions when attempting
to restart an app (in this case, the odd SecurityException when
attempting to restart Chrome) and AppInstance.restartApp() wasn't
handling it, leaving the AppScene in limbo.  Now it treats an
exception the same way as when startTask() returns false and by way
of close(), AppRing disposes of the AppScene.

However, rotateApp() has already been called and posted its
startAnimationRunnable, and it is often the case that by the time it
is getting ready to run the layout, the app is no longer in the
runningApps List.  This results in passing -1 as the index of the app
to do layout on and LayoutManager ends up throwing an out-of-range
exception which crashes VR Top.

Added code to check the index of the app we're attempting to rotate
and bailout if it's less than zero.

Also: I discovered that several classes were using GearVRF's version
of Utility.Log, which does not write to the persistent VR Top log
files.  Switched those for VR Top's local version of Utility.Log.

Cleaned up some unused import and variables while I was at it.

Change-Id: Ib7f5d97ff9f2bed40187608a0d20dffd013209c4