Fixed AppTray layout issue
A log in Widget.registerPickable() didn't have a TAG argument, so the
call was failing with a NPE, resulting in the layout call getting
bypassed.

Also removed some unnecessary code from AppTrayWidget.destroy().
When a child Widget is removed, the child's doOnDetached() gets
called which -- after setting mIsAttached to false -- calls
registerPickable(), which in turn does the clean up that the calls to
setFocusEnabled(false) and setTouchable(false) are intended to do.
Also, hiding the label directly before removing it doesn't accomplish
anything that simple removing it does.

Change-Id: I9d7ac1269f3bc76d8741f94a4eb597852d0f29f8