(f20221) [REFACTOR] RingList item host
Added a class, ListItemHostWidget, whose purpose is to hold Widgets
obtained from Adapter.getView() so that they can be positioned within
the RingList transparently; i.e., giving the hosted Widgets the
freedom to transform without concern for their position in the
RingList.  ListItemHostWidget also caches the position in the data
set and item ID for the Widget it's hosting.

Also fixed a few things that I discovered doing this work:
- Sometimes there is no EXTRA_TEXT field in the
StatusBarNotifications received by NotificationListener. NPEs could
sometimes result.
- AppTray's logging was spewing the full dump of the AppScene in
several locations; replaced this with just the name.
- AppTray.resetData() and clearCache() were doing a lot more than
simply resetting and clearing; tiny refactor to produce updateData()
and updateCache() so that the original functions are now limited to
just resetting and clearing, which makes destroy() leaner.
- AppRing was using "AppTray" instead of TAG for a couple of logs.

Patch 2: rebased

Patch 3: rebased

Change-Id: I532377d881ec18fd8acbfd206f981b1bbb50174a