(f20224) Track Widget transform changes, use transformed bounding box
Modified TransformCache to expose convenience methods for updating
its members and return whether the transform was altered. This
information will, in a later commit, be used by animations to
determine whether they should call for the affected Widget to be run
through a layout cycle.

Updated Widget to use TransformCache's tracking throughout and to
notify onTransformChanged() as appropriate.  Currently this doesn't
do much beyond invalidating the bounding box, but mChanged will be
used in a later commit that reworks the layout process.

Added BoundingBox class.  Convenient for getting depth, height, width
of a Widget's bounding box, but the main feature is that it will
transform the bounding box from GVRMesh.getBoundingBox() so that the
box is axis-aligned with the Widget's parent: depth, height, and
width are reckoned from the parent's perspective, not the child's,
and is much more useful for layout calculations when the child is
rotated, scaled, or translated.

Updated LinearLayout and RingLayout to use
Widget.getDepth/Height/Width (which refer to the corresponding
BoundingBox methods) instead of
LayoutHelpers.calculateGeometricDimensions(), which does not give
properly transformed values.

Updated RingList logging to include the RingList's name; it's much
easier to tell what's going in the logs on that way.

Also, minor modification to how sGLThread is managed.  If a Widget
tries to use runOnGlThread() before sGLThread has been initialized,
a NPE can result. This change inits sGLThread with a `null`
WeakReference to fix that.  Now, if Widget.onInitListener has not
been called (thereby properly initializing sGLThread), a Runnable
passed to runOnGlThread() will get directly posted to
GVRContext.runOnGlThread().

Patch 2: Addressed Xu's comments

TransformCache:
- Fixed || vs |
- Forwarded save(Widget, boolean) to save(GVRTransform, boolean) to
save some method invocation cycles
- Added GVRTransform version of changed() for the same reason

LinearLayout, RingLayout
- Replaced 0, 1, 2 with Axis enum
- Very minor refactor in RadialLayoutStrategy.getChildSize() to
remove code duplication by using super.

Patch 3: Fixed BoundingBox transform issues

The transform was only being done with the min and max corners in
a way that was supposed to be an optimization.  Turns out it was just
busted.  Now the entire set of vertices for the mesh's bounding box
are retained and used to do transforms.  The min and max corners are
computed from the full vertex set.

Change-Id: Ic0f33293ddabab4d96c0e2206c80f1eb53d1e0cc