(f20225) [REFACTOR] Rework Widget layout process.
This is a fairly significant change to how layout is done in the
Widget framework; the logic is more similar to the Android flow.

When a Widget's transform has changed, or a GroupWidget needs to
layout its children, requestLayout() is called, rather than calling
layout() directly.  This marks the Widget as needing layout, and
communicates the request up the scene graph until it gets to the root
(RootWidget in MainScene).  At this point, a Runnable is scheduled on
the GL thread to execute the actual layout process.  Layout then
proceeds down the scene graph.

At each level of the scene graph, each Widget has layout() called on
it.  layout() checks if 1) a layout has been explicitly requested
or 2) the Widget has been marked as "changed" because of alterations
to its transform.  If either of these is true, onLayout() is called.
onLayout() is a hook method that Widget override in order to perform
their own layout, on themselves and, potentially, on any children
they have.

This abstraction of the layout process allows us to have only one
layout pass per frame instead of (potentially) many.  For example,
currently, adding multiple children to a GroupWidget can result in
multiple calls to layout().  This scheme ensures that only one call
to layout will get run per frame.

KNOWN ISSUES:

Currently, layout is broken for both backgrounds and the app tray.
Because 6 other tickets are blocked on this change getting in, we're
going to treat these as separate bugs.

Change-Id: I86a31c3ad936f06b95d347531d40f2540395089d