(f20227) Trigger layout from animation changes
Modified Animation to check whether the target was changed (i.e., its
transform was altered) and to request a layout on the target if it
was.

The other Animation classes add something extra: because they wrap
GVRAnimation-derived classes that operate directly on the Widget's
underlying GVRTransform (instead of using Widget's transform-related
methods), the Widget doesn't actually know if its transform has been
altered.  To work around this, the other Animation classes call
checkTransformChanged() on their target, prompting it to update its
"changed" state if necessary, so that Animation's new
check-and-request code can do its job.

Client code that derives from Animation or any of its descendents
doesn't need to worry about any of this: simply use Widget's
transform methods and it just works.

Change-Id: I5c56d32af049b03c6caa813c17557075e106aea3