(b20235) AppTray and Background Picker layout broken by #20225
Per the note's in the ticket, I made minor modifications so that the
AppTray and HorizontalPicker had Widget-based connections to the
MainScene and the problems were resolved.

Also added some extra logging to help diagnose layout issues.

Also muzzled the FocusManager because there is such a thing as too
many logs.  Set FocusManager.LOGGING_ENABLED=true if you want to see
them.

Change-Id: I5b2511894cfd4cb12b02b7aaad90f235ed1dd39d