(f20238) Implement setItemFocusEnabled()
Added setItemFocusEnabled() to RingList.  This performs a blanket
disable of item focus when called with `false`; the data set is
refreshed when called with `true` so that the Adapter can update
setFocusEnabled() for its views as needed.

It's not bullet proof: Adapter's have to call setFocusEnabled() for
their items if they're going to be used with this API, otherwise
their focus may be permenantly disabled.  Also, it doesn't prevent
code outside RingList from setting focus enabled on Widget obtained
from Adapter.getView().  However, this is documented on the new
method, so devs should be able to manage this without too much
trouble.  I considered some solutions that might be more bullet
proof, but they were much bigger, hairier, and fragile.  This
solution is simple and just requires a little awareness on the part
of devs.

AppRing2 was modified to call the new method from its implementation
of enableAppFocus().  AppsAdapter.getView() was also slightly
restructured to appropriately handle item focus being re-enabled.

FocusManager was modified so that when the current focused Focusable
is unregistered, the Focusable loses focus.  Also pulled the Runnable
in onDrawFrame() out into a member variable: we really don't need to
allocate a new Runnable every single frame; just the one will do.

Change-Id: Icc87ec22985261e8abc5220eab1b4277b9974e89