(f20237,f20232) Integration: AppInstance handling and miscelleneous
This change integrates the bulk of old AppRing's code into AppRing2,
including most (if not all) of the work for 20232 (Integrate
AppInstance handling). I've split drag handling out into a separate
ticket; it shouldn't be all that difficult to do, but it's a sizeable
chunk of code, and the diff is already pretty nasty.

I've put TODO markers all over to mark where work for some of the
other tickets needs doing.

NOTE: AppRing2 is *still* not enabled by default; I don't want folks
freaking out about the nightly build.  Just remove the inline comment
in buildAppRingContentScene() to enable it.

Change-Id: I288e5c6a8d8ccb3bc9e16e49fceb003179710ed6