(b20240) AppRing2: Integrate drag handling
Integrated drag handling code from old AppRing into AppRing2.  You
can't see anything in VR Top yet, because there's a dependency on
RingList.scrollBy(float) (Redmine #20241), but if you watch the
logging output, you can see the rotation values.

There's a significant chunk of code for rearranging apps that I
brought over as well.  We currently don't support this, and it's
likely that we'd actually do most of the work in RingList, not
outside like this, but it keeps the existing code current and that
may be helpful if we bring the feature back.

Also: clarified the javadoc for RingList.scrollItemTo(). The existing
language could easily be interpreted as the method supporting
rearranging items in the ring, rather than the intended
functionality.

Also also: minor bits of tidying.

Patch 2: Rebased

Patch 3: Rebased

Change-Id: I44c5041cdb7f17a79325e2f8d96f96708953d36e