Miscellaneous fixes
Added a option to Animation to disable calling requestLayout() when
the target's transform has been modified.  This is useful for both
AppTray and HorizontalPicker (list of backgrounds in the Settings
content scene) -- we actually don't what the items in those lists
moving around when one of them alters their transform (scaling, in
both cases).

Also, happened upon a crashing bug.  Apparently, we can receive
onTaskRemoved notifications from the Virtual Screen Session even when
we haven't launched an app ourselves.  If this happens before we've
instantiated the App Ring, we get a NPE on the MainThread that will
crash the app after a bit.

Also also, renamed a method in RingList for clarity.

Patch 2: rebased

Change-Id: I38884419f4764259c651723415ac8eb8b54ab875