(b20258) Global menu hover highlighting
Changes to support hover highlighting in the Global Menu bar.

Added APIs for treating parent Widgets and their children as a single
focusable:

- setChildrenFollowFocus(): All children of this Widget are grouped
with it
- setFollowParentFocus(): This one Widget is grouped with its
parent

setChildrenFollowFocus() has a pretty hefty javadoc comment.

WidgetStateInfo leverages setFollowParentFocus(): sub-Widgets
representing a single logical Widget's state need to act in concert
with that widget in order for state changes to work correctly.

A couple of minor things piggybacking:

- Widgets are supposed to be touchable and focusable by default;
currently, however, if the metadata in an object's name doesn't
explicitly set those properties `true`, they will be set to `false`.
Modified the NodeEntry Widget constructor to correct that.
- I realized that mIsAttached is basically a synonym
for `mParent != null` that we only check in registerPickable(),
whereas the equality expression is used in many places.  Replaced
this one usage and eliminated mIsAttached.
- In addition to adding support for children-follow-focus, modified
doOnFocus() to only set mIsFocused once doFocus() has been called
and had its opportunity to reject focus.
- Moved call to updateState() from onFocus() to doOnFocus() so that
the ordering is correct.

Patch 2: Rebased

Patch 3:
- Took Srini's suggestion re. multiple loops in
setChildrenFollowFocus()
- Added parallel APIs for following the parent's input; i.e., tapping
on one Widget in the group is like tapping all of them.  Also needed
for properly handling Widget state.
- Got rid of static reference to TouchManager
- Some member and parameter renaming

Change-Id: Ib491d26cd306d04cb3cd7f91092204d872f250ac