Adding focused states for Home's Phone and Vr buttons.
Patch 5: Changes in the names of the home panel buttons made touch
events not work.  Fixed up the names expected in AppPanelController

Also, the "potential to cause other problems" mentioned for the fix
to #20321 showed up quickly: there is some fast switching between
states for the Home button when VR Top starts up.  Sometimes this
results in the button being removed from the scene graph after we've
tried to add it: no active home button shown.  Once the user has
switched to another content scene and back, it's fine and continues
to be since the user can't realistically switch between two content
scenes in the space of a single frame.  But it doesn't look good to
be missing a button at start up, even if it isn't one you need anyway
(because you're already at "home").

To fix this, both the show and hide in WidgetStateInfo.set() are now
wrapped in runInGlThread().  Removed the previous instance of this in
Widget.setVisibility(), as this issue has not caused a problem before
and this limits the hackery to the one instance it has.

Change-Id: I2ed7a858e1e4ac0de51c3356e0084b3f92ec8892