(b20263) try to open more than 8 apps, then try to close all
Now we remove a child from its current parent before the child is
added to a different parent. This prevents the child from being
kicked out of the pickable list.

Still when we say 'close all' from App Tray not all of the apps
are closed. These seem to be due to one SceneObject getting
attached to multiple widgets.

Change-Id: I233b7d5a525fd259c342c91d20b73a459ca1bbf6