(b20322) Context menu loses focus to app in background
Wrapped mWindowControlContainer in a Widget and use that Widget where
ever possible.  This allows us to use Widget's setParentFollowFocus()
so that the app retains focus even when the cursor is on the context
menu.

Didn't completely convert ControlBar to a Widget because, frankly,
that introduced issues; wrapping was the most conservative thing to
do.  We can revisit this later.

Also modified FocusManager to better detect when a Focusable should
lose focus; it allowed focus to remain in the case that there wasn't
another Focusable to take it.

Change-Id: Iaaa0233b8559e4127058e70623f4db28af9c6beb