(b20384) Back key quit maximized app leaves gap in AppRing
The key part of the change is that in AppRing2.returnBorrowedApp(),
if the app is *not* launched, we now have the adapter notify that
there's been a change so that the RingList can properly layout the
remaining items.

Also:
- BigScreenViewManager.setApp() was not handling a null AppScene
parameter properly and a NPE was getting thrown. This wasn't
contributing to the problem but it's not a good thing.
- Modified a lot of spots where we log an AppScene to just print the
AppScene's name, and not the whole schmorgasbord of internal
information because good grief there is way too much noise in the
logs.
- Subjected Widget.onTransformChanged() to the Policy.LOGGING_VERBOSE
blockade because seriously, way. Too. Much. Noise.
- Cleaned up a couple of minor bits of dead code.

Change-Id: If272ec3aae6e211a56adf149a14d0e7ab3fe43cd