(b20386) Un-maximizing app messes up layout
The zoom animation in AppRing2.hide() results in a layout request
being posted for the RingList in AppRing2; however, when going to big
screen mode, by the time the request gets handled, the borrowed app
has been scaled up to 2.5.  When we return the app, the phi values
calculated by that layout request are wrong because the angular width
of the maximized app was calculated using the big screen size.

The zoom animation doesn't need to trigger a layout request, so that
option is now disabled.

Also, slightly logging improvement for RingList.

Change-Id: I7b0b9df4d0e8cab886c25953b5c2bba9c9cef50f