(b20458) Apps not always evenly spaced out
This ended up being trickier than I thought it would. To start with,
how RingList was doing layout was subtly broken. Each subsequent
item's phi was derived from the current item's phi plus it's angular
width; however, if the subsequent item was not as wide as the current
item, the phi would be too large. To fix that, phi is advanced by one
half the current item's angular width and then one half the
subsequent item's angular width, plus padding. This logic got wrapped
into layoutItems() which most of the internal layout methods now use;
it's not entirely clear to me how it should fit into
wrapAroundCenterLayout(), but that's not crucial and can be dealt
with another time.

That was sufficient to ensure that spacing between items in a
RingList is uniform.  However, there was a new problem: when an app
was rotated into portrait, it and all the apps to the right of it
would get shifted to the left.  This is because the AppRing's
RingList was set up for LEFT_ORDER layout.  No problem, just switch
to AROUND_ITEM_AT.  That LayoutType isn't used anywhere else and,
unfortunately, was not working correctly.

The problem with AROUND_ITEM_AT stems from the fact that, unlike the
other layout types, AROUND_ITEM_AT uses the existing phi of the
specified item; the other layout types generate the phis without any
reference to existing values, so they're free to simply start from
basePhi.  When AROUND_ITEM_AT does the same thing some very odd
layouts resulted.  The fix here was to keep track of the change to
basePhi between calls to onLayout(); now, the phi of the specified
item can be adjusted by basePhiDelta and the rest of the layout can
go from there.

I discovered a couple other issues:

- setLayoutType() no longer calls requestLayout(); this was
interfering with the AROUND_ITEM_AT LayoutType when an app is added
to the ring
- scrollItemTo() wasn't checking whether the item in question was
already at the specified position, so it would just get scrolled
again; also exposed by AROUND_ITEM_AT
- Each of the internal methods handling the different layout types
started with the same boiler-plate code; factored that into
getLayoutStart(), getLayoutEnd(), and getNumItemsToDisplay()
- While factoring those methods, discovered an issue with how the end
index was being calculated: if the start index was greater than zero,
it's possible, depending on the value of mMaxItemsDisplayed, to get
an end index that is greater than the size of the items array. Fixed
that.

Change-Id: Ib4f4d81c39ae132df628cdeecca33bb5864f299e