(b20591) Context menu is not shown on focused app
There was a seperate issue in play for each of the scenarios.

In the first, we were calling enableControlContainer() in response to
onLaunched() notice.  This was added as part of the implementation
for b19384 and b19091.  The engineer responsible is no longer here,
so I couldn't ask about why this was done; however, my analysis
indicates that not only is it not necessary -- enabling and disabling
the control container should be controlled by gaining and losing
focus (or, unfortunately, as part of rotation) -- it can result in
the control container getting enabled after the app scene has already
lost focus, which is ultimately the cause in this scenario.  Fixed by
removing this onLaunched() handler.

In the second scenario, we are calling enableControlContainer() in
the OnFinish handler of the rotation animation.  This can result in
the control container being shown on a rotating app that lost focus
before its rotation finished.  This is fixed by calling
setFocusedApp(null) as soon as an app has lost visual focus, rather
than only when an app is closed.  Because of this change, the
onShown() callback now checks for mFocusedApp == null.

There's another possible issue here as well: the currently rotating
app may not be the focused app, and so can end up disabling and then
enabling the context menu on another app.  Added checks to ensure
that the rotating app is the focused app before either disabling or
enabling.

Also: moved call to enableControlContainer() to before the zoom
animation is started, which results in the context menu being
available to use sooner.

Also: marked several items in AppRing2 as private that didn't need to
be exposed.

Patch 2: Updated commit message

Patch 3: Added null check to notifyOnFocusChanged()

Patch 4: The long focus code in FocusManager didn't get updated to
use MainThread like the regular focus code did, so there was a
possibility of onLongFocus() getting called after an object had
already lost focus.  This could manifest as applications zooming into
focused position, but the context menu not showing.  Fixed by moving
the long focus code to the MainThread so that it's properly
serialized with the regular focus code.

Patch 5: Per Srini's feedback, put the onLongFocus() call to
enableControlContainer() back inside an Animation.OnFinish so that we
don't regress #20330.

Change-Id: I0b3623c4f24f4cff7ffac8668eca690d60d8c33a