(b20604) Keyboard not displayed when cursor focused on text field
Fixed the issue by allowing the VirtualKeyboardScene to participate
in Widget's "children follow focus" mechanism, which is how the
context menu is also managed.

To do that, the VirtualKeyboardScene instance has to be wrapped in a
Widget so that it can be added to the AppSceneWidget rather than the
AppScene.  This entailed the following changes:

- A Widget -- namely, the AppSceneWidget -- must be passed to
AppRingEvents.onKeyboardTargetChanged()
- AppSceneWidget now has to be a public class; AppRing2 derives an
InternalAppSceneWidget for it's own purposes so that only the minimal
interface is exposed
- AppRing2 now has its own version of AppRingEvents, because it
doesn't make any sense for AppRing to pass an AppSceneWidget to
onKeyboardTargetChanged()
- Disconnected AppRing2 from IAppRing since AppRing and AppRing2 are
not sharing the same interface anymore and IAppRing was only intended
as a bridge between the two
- Since AppRing and AppRing2 are no longer interchangeable, marked
AppRing and related code as deprecated (we can take care of removing
in a later change)
- VirtualKeyboardManager's interaction with the AppScene is
abstracted into AttachedScene to facilitate transparently using
AppScene or AppSceneWidget as appropriate.  This is necessary because
other clients of VirtualKeyboardManager and VirtualKeyboardScene are
using bare AppScene/SystemDisplayScene instances

This may sound hairy, but it was done almost entirely with Eclipse's
refactoring tools, so it should be quite safe.

It was also necessary to add a "soft unregister" to FocusManager so
that unregistering a child that is following its parent's focus
doesn't cause the parent to temporarily lose focus.  This looks like
the app is "bouncing" away from the user and back when the keyboard
is dimissed.

A nice bonus of this change is that the convoluted code we had in
place for notifying AppRing2 when the keyboard had focus in order to
prevent the "zoom out" animation from running is no longer needed and
could be removed.

Targets of opportunity:
- Added visibility specifiers in a number of places
- Removed some unused imports
- Removed unused Settings and KeyboardPosition code from
VirtualKeyboardManager
- Tweaked FocusManager logging to cut down on the noise (also disabled
it by default)

Patch 2: Rebased

Change-Id: Ie89795bbcc6fcde1be20d0e026b2424d42e50715