(b20636) Pressing Back button in Learn segment resulted in prism/crash
One of the issues here (missing items in App Ring) was caused by NaN
getting passed into RingList.scrollBy().  It appears the value was
generated (ultimately) by DragManager.pixelsPerMillisecond(): if the
event time of the passed event is the same as the previous event's
time, a divide by zero will result.  According to IEEE-754, a
floating point divide by zero gives NaN, rather than a divide-by-zero
exception.

Modified DragManager.pixelsPerMillisecond() to check for matching
event times and to set the data to 0 (zero) in that case.

Also added protection to RingList.scrollBy() to ensure that we don't
allow NaN (or infinite) values to be used.  I choose to just log the
stack trace and return instead of throwing to avoid making any
relatively minor bugs into major issues.  This way, passing NaN or
infinity is a noop instead of a crash.  We can set it to throw the
exception after release to help track down any instances where we're
passing bad values.

Patch 2: Per Mihail's suggestion, AppRing2.clearRedundantApp() now
uses MainThread.runOnMainThreadNext() to realign the restarted
application.  Also, now uses alignAppWithFront() instead of the
camera yaw; generally speaking, things should be aligned with front,
not with the camera directly.

Change-Id: Ic9531c024462c6c8534c08004a3afe1f51c6e767