(b20650) Minimize geometry going into Cardboard mode
The status bar, background, and halos are now hidden in Cardboard
mode.  Did some logging and saw that there is only one child of the
scene, which is the single quad we're displaying the stereoscopic app
on.

Included some handy statistical methods on Widget and MainScene that
I used while doing this work.

Also removed an unused method in LauncherViewManager
(PrefSaver.clear()).

Change-Id: Ib3e9a2c4d0337afaedbbd959e2217e494b77f94d