(b20683) Native crash in ExternalRendererShader::render()
The cause of this issue was that we were release()ing a closed
AppScene prior to RingList processing its removal from the data set;
this could result in a frame getting rendered on the AppScene after
release() was called.  Very timing sensitive, because not only does
the frame have to get rendered at the right time, the VirtualScreen
has to finish getting cleaned up between the release() and the
render.

Fixed this by deferring the release() call until the AppSceneWidget
is removed from the scene graph and onDetach() is called on it.  Had
to update RingList's trimming functionality so that
ListItemHostWidget removed the hosted AppSceneWidget (otherwise
onDetach() never gets called and the quad hangs around rendering!)

Also, added better handling of exceptions thrown by Adapter.getView()
in RingList.

Change-Id: I5077373cdb65fb88e36b7518f85140c04260cb8c