(b20858): [WIDGET] Modifications to support PRESSED and CHECKED states
I had intended for this change to actually do something with the
PRESSED state, but in order to do that extensive modification of
TouchManager and possibly InputWrangler would be necessary.  Given
the work that Srini is currently doing to integrate the VR I/O APIs,
that doesn't seem like a good idea.

So, this is a much smaller change that really doesn't do a lot.  I'll
file a separate ticket for managing the PRESSED state.

Change-Id: If7f7ef97d64f377d0075a36e3416d133689d89e4